+++ android_appcoins_billing/src/main/java/com/appcoins/sdk/android_appcoins_billing/CatapultAppcoinsBilling.java
@@ -3,10 +3,14 @@ package com.appcoins.sdk.android_appcoins_billing;
 import android.app.Activity;
 import android.util.Log;
 
+import com.appcoins.sdk.android_appcoins_billing.exception.IabAsyncInProgressException;
+import com.appcoins.sdk.android_appcoins_billing.helpers.IabHelper;
+import com.appcoins.sdk.android_appcoins_billing.listeners.OnIabPurchaseFinishedListener;
+import com.appcoins.sdk.android_appcoins_billing.listeners.OnIabSetupFinishedListener;
+import com.appcoins.sdk.android_appcoins_billing.listeners.OnSkuDetailsResponseListener;
 import com.appcoins.sdk.billing.AppcoinsBilling;
 
 import com.appcoins.sdk.billing.Inventory;
-import com.appcoins.sdk.billing.Purchase;
 import com.appcoins.sdk.billing.PurchasesResult;
 import com.appcoins.sdk.billing.ResponseListener;
 import com.appcoins.sdk.billing.SkuDetailsParam;
