--- android_appcoins_billing/src/main/java/com/appcoins/sdk/android_appcoins_billing/FeatureType.java
@@ -1,9 +1,0 @@
-package com.appcoins.sdk.android_appcoins_billing;
-
-public class FeatureType {
-
-    public final static String ITEM_TYPE_INAPP = "inapp";
-
-    public static final String ITEM_TYPE_SUBS = "subs";
-
-}
