--- android_appcoins_billing/src/main/java/com/appcoins/sdk/android_appcoins_billing/OnIabSetupFinishedListener.java
@@ -1,5 +1,0 @@
-package com.appcoins.sdk.android_appcoins_billing;
-
-public interface OnIabSetupFinishedListener {
-    void onIabSetupFinished(IabResult result);
-}
