+++ android_appcoins_billing/src/main/java/com/appcoins/sdk/android_appcoins_billing/helpers/PayloadHelper.java
@@ -1,4 +1,4 @@
-package com.appcoins.sdk.android_appcoins_billing;
+package com.appcoins.sdk.android_appcoins_billing.helpers;
 
 import android.net.Uri;
 import android.support.annotation.NonNull;
