+++ android_appcoins_billing/src/main/java/com/appcoins/sdk/android_appcoins_billing/listeners/OnIabPurchaseFinishedListener.java
@@ -1,5 +1,6 @@
-package com.appcoins.sdk.android_appcoins_billing;
+package com.appcoins.sdk.android_appcoins_billing.listeners;
 
+import com.appcoins.sdk.android_appcoins_billing.types.IabResult;
 import com.appcoins.sdk.billing.Purchase;
 import com.appcoins.sdk.billing.ResponseListener;
