+++ android_appcoins_billing/src/main/java/com/appcoins/sdk/android_appcoins_billing/listeners/OnIabSetupFinishedListener.java
@@ -1,0 +1,7 @@
+package com.appcoins.sdk.android_appcoins_billing.listeners;
+
+import com.appcoins.sdk.android_appcoins_billing.types.IabResult;
+
+public interface OnIabSetupFinishedListener {
+    void onIabSetupFinished(IabResult result);
+}
