+++ android_appcoins_billing/src/main/java/com/appcoins/sdk/android_appcoins_billing/types/SkuType.java
@@ -1,4 +1,4 @@
-package com.appcoins.sdk.android_appcoins_billing;
+package com.appcoins.sdk.android_appcoins_billing.types;
 
 public class SkuType{
